requirements.txt
apache-airflow-providers-apache-spark
apache-airflow-providers-apache-kafka
confluent-kafka == 2.8.0
pyspark == 3.5.4
faker == 33.3.1
python-dotenv == 1.0.1
astronomer-cosmos